+++ arrivalboard/aircraft/models.py
